Add Broadcast-/MOTD-Message functionality (XEP-0133)
Adds the ability to send Announcements and to set the Message of the
Day. After Login service discovery is used to detmine support and
permission. If the user has permission a submenu will be added to the
controlbox.

When the user clicks a command, a chatbox opens with the form the server
send. The user has to fill out the form and send the command back to the
server.

Furthermore adds a "-- send from <jid>" at the end of the announcement.
THis helps users to contact the sender.